moved common libs under common/ and removed 'Configuration' from property classes
