+++ common/pom.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<modelVersion>4.0.0</modelVersion>
+	<groupId>org.springframework.cloud.stream.module</groupId>
+	<artifactId>spring-cloud-stream-modules-common</artifactId>
+	<version>1.0.0.BUILD-SNAPSHOT</version>
+	<packaging>pom</packaging>
+	<modules>
+		<module>spring-cloud-stream-modules-common-configuration</module>
+		<module>spring-cloud-stream-modules-test-support</module>
+	</modules>
+</project>
