+++ common/spring-cloud-stream-modules-test-support/src/main/java/org/springframework/cloud/stream/modules/test/PropertiesInitializer.java
