--- spring-cloud-stream-modules-common-configuration/src/main/java/org/springframework/cloud/stream/module/MaxMessagesConfigurationProperties.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2015 the original author or authors.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.stream.module;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/**
- * @author David Turanski
- */
-@ConfigurationProperties
-public class MaxMessagesConfigurationProperties {
-
-	/**
-	 * Maximum messages per poll. Default is -1 meaning no maximum.
-	 */
-	long maxMessages = -1;
-
-	public long getMaxMessages() {
-		return this.maxMessages;
-	}
-
-	public void setMaxMessages(long maxMessages) {
-		this.maxMessages = maxMessages;
-	}
-
-}
